relation mapping of two models works as intended

+++ RoomDatabaseExample/app/src/main/java/de/test/roomdatabaseexample/Note.java
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey;
 public class Note implements Cloneable{
 
     @PrimaryKey(autoGenerate = true)
-    private int noteId;
+    private long noteId;
 
     private String title;
 
@@ -25,7 +25,7 @@ public class Note implements Cloneable{
         this.noteId = noteId;
     }
 
-    public int getNoteId() {
+    public long getNoteId() {
         return noteId;
     }
